I am aprox 188cm or EC, lots of changes.
All the request changes have been made to some degree.

Notes:
 - New assertions emailaddress!=recoveryemail, lastlogin>=datecreated,
password!=passhint
 - country field should be validated through some module, but currently
there are no limitations to how they might be spelled or capitalized.
Some enum implementation might be prefered or otherwise referenced to
some standardized code.
 - phone [regex](https://regexpattern.com/phone-number/) has been
modified slightly to include all the variations found at this source.
This should also be handled externally to recieve updates and a full
version might want to assert that the number variation matches the
user's country of origin.
 - email field [regex](https://regexpattern.com/email-address/) should
be implemented with a third party service to verify instead of a raw and
static regular expression.
 - Database has been updated with some test values, they're not very
random so it's not a decent represenation.
 - There is some minor driver code for expected but not stated behavior.
Namely that only active accounts can log in.
 - Although the specifications request that passhint should be returned,
we should really treat this like a password and have seperate methods to
ensure conditional access control as otherwise we risk compromising
accounts.